SSBD201903-76 Fixes in entities (id type to object type Long), moved directories for HTML files, added Ozark with Thymeleaf Engine to the project, created registration form, added login and registration controller to the application, registration almost ready, TODO: exception handling, controller-side validation, right redirects and tests for controllers

+++ src/main/java/pl/lodz/p/it/ssbd2019/ssbd03/accountsmodule/repository/AccountRepositoryLocal.java
@@ -6,5 +6,5 @@ import pl.lodz.p.it.ssbd2019.ssbd03.repository.CruRepository;
 import javax.ejb.Local;
 
 @Local
-public interface AccountRepositoryLocal extends CruRepository<Account, Integer> {
+public interface AccountRepositoryLocal extends CruRepository<Account, Long> {
 }
