+++ src/main/java/pl/lodz/p/it/ssbd2019/ssbd03/accountsmodule/repository/UserRepositoryLocal.java
@@ -1,11 +1,10 @@
 package pl.lodz.p.it.ssbd2019.ssbd03.accountsmodule.repository;
 
-import pl.lodz.p.it.ssbd2019.ssbd03.entities.Account;
 import pl.lodz.p.it.ssbd2019.ssbd03.entities.User;
 import pl.lodz.p.it.ssbd2019.ssbd03.repository.CruRepository;
 
 import javax.ejb.Local;
 
 @Local
-public interface UserRepositoryLocal extends CruRepository<User, Integer> {
+public interface UserRepositoryLocal extends CruRepository<User, Long> {
 }
